refactor(rev57lp): move default configuration to `config/`
Looks like it is impossible to compile the firmware when the default
shield configuration files are in `boards/shields/rev57lp`. Indeed, the
keyboard isn't found and the default list is displayed.

A workaround was to give the path to the root directory to the `ZMK_CONFIG`
CMake option, instead of the absolute path to `config/`. However, this
makethings confusing.

In the end to simplify things until I managed to move the stuff into the
better place `boards/` is, I'll stick with that.